pyproject.toml
[tool.poetry]
name = "ASCMonitor"
version = "0.1.0"
description = ""
authors = ["Felix Peppert <felix.peppert@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
mendeley = "^0.3.2"
requests = "^2.5"
sqlalchemy = "^1.3"
flask_sqlalchemy = "^2.3"
mysqlclient = "^1.4"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
